utils/**Utils: Extract license file patterns into a dedicated file
To prepare for adding related logic in an upcoming change.
