+++ utils/src/main/kotlin/Utils.kt
@@ -28,31 +28,6 @@ import java.security.Permission
 
 import kotlin.reflect.full.memberProperties
 
-private fun List<String>.generateCapitalizationVariants() = flatMap { listOf(it, it.toUpperCase(), it.capitalize()) }
-
-/**
- * A list of globs that match default license file names.
- */
-val LICENSE_FILENAMES = listOf(
-    "license*",
-    "licence*",
-    "*.license",
-    "*.licence",
-    "unlicense",
-    "unlicence",
-    "copying*",
-    "copyright",
-    "patents"
-).generateCapitalizationVariants()
-
-/**
- * A list of globs that match files that often define the root license of a project, but are no license files and are
- * therefore not contained in [LICENSE_FILENAMES].
- */
-val ROOT_LICENSE_FILENAMES = listOf(
-    "readme*"
-).generateCapitalizationVariants()
-
 /**
  * The directory to store ORT (read-only) configuration in.
  */
